+++ src/main/java/com/shinhanDS5gi/memento/service/ReviewServiceImpl.java
@@ -13,9 +13,8 @@ import com.shinhanDS5gi.memento.dto.MentoReviewsListResponse;
 import com.shinhanDS5gi.memento.dto.MentoReviewsSliceResponse;
 import com.shinhanDS5gi.memento.repository.mentos.MentosRepository;
 import com.shinhanDS5gi.memento.repository.ReservationRepository;
-import com.shinhanDS5gi.memento.repository.ReviewRepository;
+import com.shinhanDS5gi.memento.repository.Review.ReviewRepository;
 import com.shinhanDS5gi.memento.repository.member.MemberRepository;
-import com.shinhanDS5gi.memento.repository.mentos.MentosRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
